Undo a commit after some crying

+++ app/src/main/java/com/example/bookbank/activities/AddBookActivity.java
@@ -84,7 +84,7 @@ public class AddBookActivity extends AppCompatActivity {
                         id,
                         title.getText().toString(),
                         author.getText().toString(),
-                        Integer.parseInt(isbn.getText().toString()),
+                        Long.parseLong(isbn.getText().toString()),
                         description.getText().toString(),
                         "Available",
                         ownerId,
